+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/questions.test.ts
@@ -142,7 +142,7 @@ describe('questions', () => {
               },
               {
                 "guiOptions": {
-                  "hint": "Select a local file that defines the service connection for an ABAP Environment on SAP Business Technology Platform",
+                  "hint": "Select a local file that defines the service connection for an ABAP Environment on SAP Business Technology Platform.",
                   "mandatory": true,
                 },
                 "guiType": "file-browser",