+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/registry.ts
@@ -80,8 +80,14 @@ export default class FEV2QuickActionRegistry extends QuickActionDefinitionRegist
                     title: 'ANALYTICAL LIST PAGE',
                     definitions: [
                         AddControllerToPageQuickAction,
+                        AddPageActionQuickAction,
+                        ToggleClearFilterBarQuickAction,
+                        ToggleSemanticDateRangeFilterBar,
                         ChangeTableColumnsQuickAction,
-                        AddTableCustomColumnQuickAction
+                        AddTableActionQuickAction,
+                        AddTableCustomColumnQuickAction,
+                        EnableTableFilteringQuickAction,
+                        AddNewAnnotationFile
                     ],
                     view,
                     key: name + index