+++ packages/flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/flp-config-sub-generator
 
+## 0.0.16
+
+### Patch Changes
+
+-   6b6c64a: adds new module cf deploy config sub gen
+-   Updated dependencies [6b6c64a]
+    -   @sap-ux/deploy-config-generator-shared@0.0.11
+
+## 0.0.15
+
+### Patch Changes
+
+-   @sap-ux/flp-config-inquirer@0.2.16
+-   @sap-ux/app-config-writer@0.5.15
+
 ## 0.0.14
 
 ### Patch Changes